+++ lib/filter_list.dart
@@ -1,4 +1,5 @@
 library filter_list;
 
 export 'src/filter_list_dialog.dart';
+export 'src/filter_list_delegate.dart';
 export 'src/theme/theme.dart';