+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   - Removed `wrapCrossAxisAlignment` option
   - Removed `wrapSpacing` option
   - Removed `borderRadius` option
+  - Replace `ItemSearchDelegate<T>` with `SearchPredict<T>` method
 
 - Add `FilterListTheme` for filter list widget theme
 - Add `ChoiceChipTheme` for choice chip theme.
@@ -31,6 +32,7 @@
 - Add `ControlButtonBarTheme` for control button bar theme
 - Add `ControlButtonTheme` for control button theme
 - Add `controlButtons` option to display/hide control buttons (All, Reset)
+
 
 ## [1.0.0] 21 May 2021